+++ src/preset/node/types.ts
@@ -4,7 +4,7 @@ import type { Request, Response, NextFunction } from 'express';
  * Контекст обработчика express.
  */
 export interface ExpressHandlerContext {
-  req: Request;
-  res: Response;
-  next: NextFunction;
+  readonly req: Request;
+  readonly res: Response;
+  readonly next: NextFunction;
 }
